Add utility to compute folder path from import.meta.url
The code was refactored to add a utility function that computes the folder path of the current JavaScript module file. This function is needed because the traditional __dirname global of Node.js CommonJS modules is not available in ES6 modules. The function accepts the special import.meta.url of ES6 modules and returns the local directory path that corresponds to it. The function has been utilized in the helper.js, botbase-factory.js, botbase.test.js and helper.test.js files. Also, the writeIPToFile method in helper.js was updated to accept full file path instead of base path.